Update index.html
Update year to 1996 because it's a better year 😎 
It's the year one of my favorite games came out: "Crash Bandicoot".

Small joke, feel free to close the pull request. Just wanted to say it is an awesome idea to have a fully and completely working offline gaming device.

Appreciate you sharing this! 🚀